requirements.txt
# Core runtime dependencies
pydantic>=2.8,<3.0
PyYAML>=6.0
